requirements.txt
Flask
Flask-Bootstrap
Flask-CLI
Flask-SQLAlchemy
Flask-WTF
Pillow
PyMySQL
cryptography
pyzbar
